# cooperation/src/main/resources/static/doc/openapi_v2.yaml
openapi: 3.0.1
info:
  title: Etablerad samverkan i vården
  description: "![Informationsmodell - Cooperation](images/cooperation.png)\r\n\r\nInformation om tjänster, producenter och konsumenter baserad på tjänsteadresseringskatalogen i både Ineras nationella tjänsteplattform och regionala tjänsteplattformar. Eventuella framtida konfigurationsändringar syns inte i api:et, enbart hur det ser ut i dagsläget. Konfiguration i en tjänsteplattform kan ibland ske på en viss nivå i HSA-trädet, och avser alla noder därunder. Detta speglas rakt av i api:et och HSA-trädet evalueras inte här. De tekniska id-begreppen på alla objekt genereras om en gång per dygn och garanteras således inte ha samma värde två dagar i rad. Applikationer bör skrivas så att de är oberoende av de tekniska nycklarna. Varje resurs har sin egen nummerserie för de tekniska id-begreppen. En api-nyckel kan behövas för att nå api:et."
  version: 2.0.0
servers:
  - url: /coop/api/v2
paths:
...
  /serviceProducers:
    get:
      summary: Service Producer - tjänsteproducent
      description: |
        Alla tjänsteproducenter. En producent identifieras verksamhetsmässigt av hsa-id. En förekomster av detta objet representerar en producent i en instans av tjänsteplattformen. Innebär att samma hsa-id kan förekomma på två objekt. Om connectionPoint inte anges som query-parameter så inkluderas connectionPoint till varje objekt i svaret 
      parameters:
        - name: hsaId
          in: query
          description: Sök bara tjänsteproducenter som har angivet hsaId.
          required: false
          schema:
            type: string
        - name: connectionPointId
          in: query
          description: Filtrerar bara tjänsteproducenter som hör till den aktulla instansen av tjänsteplattformen. Ange den tekniska nyckeln för connectionPoint tex 376.
          required: false
          schema:
            type: integer
            format: int64
        - name: logicalAddressId
          in: query
          description: Filtrerar bara tjänsteproducenter som publicerar någon tjänst med en specifik logisk adress. Ange den tekniska nyckeln för LogicalAddress tex 42.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceContractId
          in: query
          description: Filtrerar bara tjänsteproducenter som publicerar någon tjänst med ett specifik tjänstekontrakt. Ange den tekniska nyckeln för ServiceContract tex 123.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceConsumerId
          in: query
          description: Filtrerar bara tjänsteproducenter som har någon samverkan med en specifik tjänstekonsument. Sökningen utförs genom att titta på tjänstproducenter som publicerar en tjänst på en viss logisk adress samtidigt som någon tjänstekonsument är konfigurerad att anropa samma logiska adress.  Ange den tekniska nyckeln för ServiceConsumer tex 538.
          required: false
          schema:
            type: integer
            format: int64
      tags:
        - ServiceProducer
      responses:
        '200':
          description: En lista av tjänsteproducenter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceProducer'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceProducer'
                xml:
                  wrapped: true
                  name: serviceProducers
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /serviceProducers/{id}:
    get:
      summary: Service Producer - tjänsteproducent
      description: |
        En specifik tjänsteproducent.
      parameters:
        - $ref: '#/components/parameters/TechnicalKeyParam'
      tags:
        - ServiceProducer
      responses:
        '200':
          description: En tjänsteproducent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProducer'
            application/xml:
              schema:
                $ref: '#/components/schemas/ServiceProducer'
        '404':
          description: ServiceProducer not found
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /serviceProductions:
    get:
      summary: ServiceProduction - tjänsteproduktion
      description: |
        Alla tjänsteproduktioner. En tjänsteproducent är konfigurerad att publicera en tjänst via en logisk adress avseende ett tjänstekontrakt. Genom att använda include-parametern hängs det på information från 1-4 relaterade objekt som beskriver tjänsteproduktionen. OBS prova inte denna resurs på denna sida utan att ange minst en filtrerings-parameter. Det är ganska stor träffbild om man får alla förekomster tillbaka och de skall parsas i dokumentverktyget, tar lång tid.  
      parameters:
        - name: physicalAddress
          in: query
          description: Sök bara tjänsteproduktioner som har angiven fysisk adress.
          required: false
          schema:
            type: string
        - name: rivtaProfile
          in: query
          description: Sök bara tjänsteproduktioner som har angivet rivTaProfil.
          required: false
          schema:
            type: string
        - name: connectionPointId
          in: query
          description: Filtrerar bara tjänsteproduktioner som är konfigurerade i den aktulla instansen av tjänsteplattformen. Ange den tekniska nyckeln för connectionPoint tex 376.
          required: false
          schema:
            type: integer
            format: int64
        - name: logicalAddressId
          in: query
          description: Filtrerar bara tjänsteproduktioner som avser en specifik logisk adress. Ange den tekniska nyckeln för LogicalAddress tex 42.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceContractId
          in: query
          description: Filtrerar bara tjänsteproduktioner som avser ett specifik tjänstekontrakt. Ange den tekniska nyckeln för ServiceContract tex 123.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceProducerId
          in: query
          description: Filtrerar bara tjänsteproduktioner avseende en specifik tjänsteproducent.  Ange den tekniska nyckeln för ServiceProducer tex 538.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceDomainId
          in: query
          description: Filtrerar bara samverkan avseende alla kontrakt som ingår i en viss tjänstedomän.  Ange den tekniska nyckeln för ServiceDomain tex 825.
          required: false
          schema:
            type: integer
            format: int64
        - name: include
          in: query
          description: 'Styr hur mycket information som skall skickas tillbaka i svaret.  Ange en eller flera av följande objekt kommaseparerade: connectionPoint,logicalAddress,serviceContract,serviceProducer'
          required: false
          schema:
            type: string
            default: connectionPoint,logicalAddress,serviceContract,serviceProducer
      tags:
        - ServiceProduction
      responses:
        '200':
          description: En lista av tjänsteproduktioner
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceProduction'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceProduction'
                xml:
                  wrapped: true
                  name: serviceProductions
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /serviceProductions/{id}:
    get:
      summary: ServiceProduction - tjänsteproduktion
      description: |
        En specifik tjänsteproduktion. Alla 4 relaterade objekt är alltid inkluderade.
      parameters:
        - $ref: '#/components/parameters/TechnicalKeyParam'
      tags:
        - ServiceProduction
      responses:
        '200':
          description: En tjänsteproduktion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProduction'
            application/xml:
              schema:
                $ref: '#/components/schemas/ServiceProduction'
        '404':
          description: ServiceProduction not found
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
...
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
  responses:
    UnexpectedErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    TechnicalKeyParam:
      name: id
      in: path
      description: Teknisk nyckel
      required: true
      schema:
        type: number
  schemas:
    ConnectionPoint:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        platform:
          type: string
          description: Plattform tex NTJP.
        environment:
          type: string
          description: Miljö tex PRODUCTION.
        snapshotTime:
          type: string
          description: Den tidpunkt som informationen lästes av från tjänsteplattformen.
      xml:
        name: connectionPoint
...
    LogicalAddress:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        logicalAddress:
          type: string
          description: logisk adress.
        description:
          type: string
          description: Beskrivning.
      xml:
        name: logicalAddress
    ServiceContract:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        name:
          type: string
          description: Namn.
        namespace:
          type: string
          description: Namnrymd.
        major:
          type: number
          description: Major-version exvis 1 eller 2.
        minor:
          type: number
          description: Minor-version exvis 0 eller 1.
      xml:
        name: serviceContract
...
    ServiceProducer:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        hsaId:
          type: string
          description: hsa-id.
        description:
          type: string
          description: Beskrivning.
      xml:
        name: serviceProducer
    ServiceProduction:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        physicalAddress:
          type: string
          description: fysisk adress för tjänsten.
        rivtaProfile:
          type: string
          description: Den profil som tjänsten stöder.
        serviceProducer:
          $ref: '#/components/schemas/ServiceProducer'
        logicalAdress:
          $ref: '#/components/schemas/LogicalAddress'
        serviceContract:
          $ref: '#/components/schemas/ServiceContract'
        connectionPoint:
          $ref: '#/components/schemas/ConnectionPoint'
      xml:
        name: serviceProduction
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        fields:
          type: string
      xml:
        name: error